Move existing filtering tests to separate test class
